Diversas melhorias, a fase_2 consegue instânciar aleatóriamente as plataformas e também, os inimigos que aparecem nelas. Quase finalizada, falta apenas setar a movimentação do Capitão e terminar o obstaculo dificil que causa dano.

+++ Jogo_TecProg_2024/Source/Entidades/Obstaculos/Esteira.cpp
@@ -2,10 +2,10 @@
 
 // Implementações usando qualificação completa de namespace
 Entidades::Obstaculos::Esteira::Esteira(float vel, float dir) :
-Obstaculo(), velocidadeTransporte(vel), direcao(dir) {
+Obstaculo(), velocidadeTransporte(vel), direcao(static_cast<int>(dir)) {
 
     agressivo = false;
-        tamanhoEsteira = sf::Vector2f(100.0f, 10.0f); // Exemplo de tamanho
+    tamanhoEsteira = sf::Vector2f(100.0f, 10.0f); // Exemplo de tamanho
 
 }
